server/requirements.txt
flask>=3.0
flask-cors>=4.0
python-docx>=1.1
Jinja2>=3.1
playwright>=1.40
Pygments>=2.17
python-dotenv>=1.0
requests>=2.31
PyYAML>=6.0
Pillow>=10.0
